refactor: Ut 클래스 분리
